pyproject.toml
[tool.poetry]
name = "repl_python3_Split"
version = "0.1.0"
description = ""
authors = ["yujinyang1998 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pil-compat = "^1.0.0"
pytesseract = "^0.3.8"
telegram = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
